# openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "GreyhoundIQ HTTP API",
    "description": "Source-derived contract for GreyhoundIQ Next.js route handlers. Only the local server is listed so automated security testing cannot accidentally target production. Operations marked with x-authentication-status: unknown require manual security-contract verification before release.",
    "version": "0.1.0",
    "contact": {
      "name": "GreyhoundIQ Support",
      "url": "https://greyhoundsiq.com.au/contact",
      "email": "support@greyhoundsiq.com.au"
    }
  },
  "servers": [
    {
      "url": "https://greyhoundsiq.com.au",
      "description": "GreyhoundIQ production edge",
      "x-internal": false
    }
  ],
...
  "paths": {
...
    "/api/dogs/search": {
      "get": {
        "operationId": "getDogsSearch",
        "summary": "Get /api/dogs/search",
        "description": "Source-derived contract for `GET /api/dogs/search`. Unspecified object shapes remain intentionally open until their source schemas are registered.",
        "tags": [
          "Dogs"
        ],
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "maxLength": 256
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UnspecifiedJsonObject"
                }
              }
            }
          },
          "429": {
            "$ref": "#/components/responses/RateLimitExceeded"
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UnspecifiedJsonObject"
                }
              }
            }
          }
        },
        "security": [],
        "x-endpoint-id": "HTTP.GET.API_DOGS_SEARCH",
        "x-source-file": "src/app/api/dogs/search/route.ts",
        "x-authentication-status": "public",
        "x-registry-verification": "Not verified",
        "x-request-schema-source": "not verified",
        "x-output-schema-source": "not verified",
        "x-known-gaps": [
          "Discovered from source; security metadata and implementation trace have not been verified."
        ],
        "x-rate-limit-policy": "RATE.API.getDogsSearch"
      }
    },
...
    "/api/reports": {
      "get": {
        "operationId": "getReports",
        "summary": "Get /api/reports",
        "description": "Source-derived contract for `GET /api/reports`. Unspecified object shapes remain intentionally open until their source schemas are registered.",
        "tags": [
          "Reports"
        ],
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "maxLength": 2048
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UnspecifiedJsonObject"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UnspecifiedJsonObject"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UnspecifiedJsonObject"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UnspecifiedJsonObject"
                }
              }
            }
          }
        },
        "security": [
          {
            "WorkOSSession": []
          }
        ],
        "x-endpoint-id": "HTTP.GET.API_REPORTS",
        "x-source-file": "src/app/api/reports/route.ts",
        "x-authentication-status": "required",
        "x-registry-verification": "Not verified",
        "x-request-schema-source": "not verified",
        "x-output-schema-source": "not verified",
        "x-known-gaps": [
          "Discovered from source; security metadata and implementation trace have not been verified."
        ]
      },
      "post": {
        "operationId": "postReports",
        "summary": "Post /api/reports",
        "description": "Source-derived contract for `POST /api/reports`. Unspecified object shapes remain intentionally open until their source schemas are registered.",
        "tags": [
          "Reports"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UnspecifiedJsonObject"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UnspecifiedJsonObject"
                }
              }
            }
          },
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UnspecifiedJsonObject"
                }
              }
            }
          },
          "400": {
            "description": "Bad request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UnspecifiedJsonObject"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UnspecifiedJsonObject"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UnspecifiedJsonObject"
                }
              }
            }
          },
          "429": {
            "$ref": "#/components/responses/RateLimitExceeded"
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UnspecifiedJsonObject"
                }
              }
            }
          }
        },
        "security": [
          {
            "WorkOSSession": []
          }
        ],
        "x-endpoint-id": "HTTP.POST.API_REPORTS",
        "x-source-file": "src/app/api/reports/route.ts",
        "x-authentication-status": "required",
        "x-registry-verification": "Not verified",
        "x-request-schema-source": "not verified",
        "x-output-schema-source": "not verified",
        "x-known-gaps": [
          "Discovered from source; security metadata and implementation trace have not been verified."
        ],
        "x-rate-limit-policy": "RATE.API.postReports"
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "WorkOSSession": {
        "type": "apiKey",
        "in": "cookie",
        "name": "wos-session",
        "description": "WorkOS AuthKit session cookie; the name is source-configurable through WORKOS_COOKIE_NAME."
      },
      "OnboardingAnalyticsConsent": {
        "type": "apiKey",
        "in": "header",
        "name": "x-greyhoundiq-analytics-consent",
        "description": "Required fixed consent assertion for the privacy-minimised public analytics write. This is a request guard, not a user credential or authorization grant."
      },
      "InternalSecret": {
        "type": "apiKey",
        "in": "header",
        "name": "x-internal-secret"
      },
      "InternalBearer": {
        "type": "http",
        "scheme": "bearer",
        "description": "Alternative internal bearer secret accepted by requireInternalRequest."
      },
      "StripeSignature": {
        "type": "apiKey",
        "in": "header",
        "name": "stripe-signature"
      },
      "LagoSignature": {
        "type": "apiKey",
        "in": "header",
        "name": "x-lago-signature"
      },
      "LiveKitWebhookAuthorization": {
        "type": "http",
        "scheme": "bearer",
        "description": "LiveKit webhook Authorization signature verified by livekit-server-sdk."
      },
      "ReplaySignature": {
        "type": "apiKey",
        "in": "query",
...
    "schemas": {
      "UnspecifiedJsonObject": {
        "type": "object",
        "additionalProperties": true,
        "description": "JSON object whose exact field contract is not yet source-registered; no fields are invented."
      },
...
    },
    "responses": {
      "RateLimitExceeded": {
        "description": "The caller exceeded the operation's bounded request policy.",
        "headers": {
          "Access-Control-Allow-Origin": {
            "description": "The exact production web origin allowed to read this response.",
            "schema": {
              "type": "string",
              "enum": [
                "https://greyhoundsiq.com.au"
              ]
            }
          },
          "Cache-Control": {
            "description": "Prevents shared or browser caches from storing the caller-specific denial.",
            "schema": {
              "type": "string",
              "enum": [
                "private, no-store"
              ]
            }
          },
          "RateLimit-Limit": {
            "description": "Maximum requests allowed in the active fixed window.",
            "schema": {
              "type": "integer",
              "format": "int32",
              "minimum": 1,
              "maximum": 2147483647
            }
          },
          "RateLimit-Remaining": {
            "description": "Requests remaining in the active window, clamped to zero through the configured limit.",
            "schema": {
              "type": "integer",
              "format": "int32",
              "minimum": 0,
              "maximum": 2147483647
            }
          },
          "RateLimit-Reset": {
            "description": "Whole delta-seconds until the active window resets; never less than one.",
            "schema": {
              "type": "integer",
              "format": "int32",
              "minimum": 1,
              "maximum": 2147483647
            }
          },
          "Retry-After": {
            "description": "Whole delta-seconds before the caller should retry; equal to RateLimit-Reset.",
            "schema": {
              "type": "integer",
              "format": "int32",
              "minimum": 1,
              "maximum": 2147483647
            }
          }
        },
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "additionalProperties": false,
              "required": [
                "error"
              ],
              "properties": {
                "error": {
                  "type": "object",
                  "additionalProperties": false,
                  "required": [
                    "code",
                    "message"
                  ],
                  "properties": {
                    "code": {
                      "type": "string",
                      "minLength": 1,
                      "maxLength": 128,
                      "pattern": "^[a-z][a-z0-9_.-]{0,127}$"
                    },
                    "message": {
                      "type": "string",
                      "minLength": 1,
                      "maxLength": 256,
                      "pattern": "^[^\\r\\n]{1,256}$"
                    }
                  }
                }
              }
            }
          }
        }
      }
    }
  },
...
}